[REFACTOR] video remove api 성능 개선

+++ src/main/java/com/vivid/apiserver/domain/video/application/VideoService.java
@@ -1,9 +1,12 @@
 package com.vivid.apiserver.domain.video.application;
 
 import com.vivid.apiserver.domain.individual_video.application.command.IndividualVideoCommandService;
+import com.vivid.apiserver.domain.user.application.CurrentUserService;
+import com.vivid.apiserver.domain.user.domain.User;
 import com.vivid.apiserver.domain.video.application.command.VideoCommandService;
 import com.vivid.apiserver.domain.video.application.query.VideoQueryService;
 import com.vivid.apiserver.domain.video.domain.Video;
+import com.vivid.apiserver.domain.video_space.application.VideoSpaceValidateService;
 import com.vivid.apiserver.global.infra.storage.AwsS3Service;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +22,9 @@ public class VideoService {
 
     private final VideoCommandService videoCommandService;
     private final IndividualVideoCommandService individualVideoCommandService;
+
+    private final CurrentUserService currentUserService;
+    private final VideoSpaceValidateService videoSpaceValidateService;
     private final AwsS3Service awsS3Service;
 
     /**
@@ -26,10 +32,14 @@ public class VideoService {
      */
     public void delete(Long videoId) {
 
-        Video video = videoQueryService.findById(videoId);
+        User currentUser = currentUserService.getCurrentUser();
+        Video video = videoQueryService.findWithVideoSpaceAndIndividualVideosById(videoId);
 
-        videoCommandService.delete(video);
+        videoSpaceValidateService.checkHostUserAccess(video.getVideoSpace(), currentUser.getEmail());
+
         individualVideoCommandService.deleteAll(video.getIndividualVideos());
+        videoCommandService.delete(video);
+
     }
 
     /**
